backend/requirements.txt
paddlepaddle
paddleocr
flask
flask-cors
pillow
numpy
